include domain models

+++ app/src/main/java/com/agah/furkan/androidplayground/data/web/model/request/AddProductToCartBody.kt
@@ -6,9 +6,9 @@ import com.squareup.moshi.JsonClass
 @JsonClass(generateAdapter = true)
 data class AddProductToCartBody(
     @Json(name = "productId")
-    val productId: Int,
+    val productId: Long,
     @Json(name = "userId")
-    val userId: Int,
+    val userId: Long,
     @Json(name = "discount")
     val discount: Double,
     @Json(name = "picture")
